used restsharp build uri functionality to build resource
