+++ oanda_dotnet/model/EndpointRequest.cs
@@ -1,4 +1,5 @@
-﻿using RestSharp;
+﻿using oanda_dotnet.serialization;
+using RestSharp;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -12,7 +13,6 @@ namespace oanda_dotnet.model
     public abstract class EndpointRequest
     {
         public abstract string Endpoint { get; }
-        public abstract Method Method { get; }
 
         public bool IsValid()
         {
@@ -34,5 +34,36 @@ namespace oanda_dotnet.model
         protected static string GetEnumValue<T>(T enumMember)
             where T : Enum
                 => enumMember.GetType().GetMember(enumMember.ToString()).FirstOrDefault().GetCustomAttribute<EnumMemberAttribute>().Value;
+
+        public abstract Method Method { get; }
+
+        public IRestRequest GenerateRestRequest()
+        {
+            RestRequest restRequest = new RestRequest(this.Endpoint, this.Method);
+            restRequest.JsonSerializer = NewtonsoftJsonSerializer.Default;
+
+            this.GetType().GetProperties()
+                .Where(property =>
+                    Attribute.IsDefined(property, typeof(EndpointParameterAttribute)) && //where is a request parameter
+                        property.GetValue(this) != null)
+
+                .ToList().ForEach(property =>
+                {
+                    EndpointParameterAttribute requestParameterAttribute = property.GetCustomAttribute<EndpointParameterAttribute>();
+                    var value = property.GetValue(this);
+                    if (value.GetType().IsEnum) { value = GetEnumValue((Enum)value); };
+                    if (value.GetType().GetInterface(nameof(ICollection)) != null) { value = string.Join(",", ConvertDynamicToString((ICollection)value)); }
+                    if (requestParameterAttribute.Type == ParameterType.RequestBody)
+                    {
+                        restRequest.AddJsonBody(value);
+                    }
+                    else
+                    {
+                        restRequest.AddParameter(requestParameterAttribute.Name, value, requestParameterAttribute.Type);
+                    }
+                });
+
+            return restRequest;
+        }
     }
 }
