+++ oanda_dotnet/model/StreamEndpointRequest.cs
@@ -1,15 +1,4 @@
-﻿using RestSharp;
-
-namespace oanda_dotnet.model
+﻿namespace oanda_dotnet.model
 {
-    public abstract class StreamEndpointRequest<T>
-    {
-        public abstract string Endpoint { get; }
-        public abstract Method Method { get; }
-
-        public string Resource()
-        {
-            return string.Empty;
-        }
-    }
+    public abstract class StreamEndpointRequest<T> : EndpointRequest { }
 }
